+++ pose_net/index.html
@@ -6,6 +6,7 @@
     <script src="https://unpkg.com/@tensorflow/tfjs"></script>
     <!-- Load Posenet -->
     <script src="https://unpkg.com/@tensorflow-models/posenet"></script>
+    <script src="https://unpkg.com/tone"></script>
     <script src="sketch.js"></script>
   </head>
   <body></body>
