--- requirements.txt
@@ -1,10 +1,0 @@
-pandas
-tensorflow
-Flask
-flask-cors
-joblib
-sklearn
-jupyterlab
-matplotlib
-h5py
-
